# apps/ai-engine/src/main.py
"""Legal AI Engine - FastAPI Service.

This service provides AI-powered legal assistance including:
- Document generation
- Legal Q&A
- Case law search
"""

import uuid
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware

from .models.requests import (
    AskQuestionRequest,
    GenerateDocumentRequest,
    SearchRulingsRequest,
)
from .models.responses import (
    AnswerResponse,
    Citation,
    DocumentGenerationStatus,
    GenerateDocumentResponse,
    Ruling,
    SearchRulingsResponse,
)

app = FastAPI(
    title="Legal AI Engine",
    description="AI-powered legal assistance platform",
    version="0.1.0",
    docs_url="/docs",
    redoc_url="/redoc",
)

# CORS middleware for development
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://localhost:3001"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# In-memory storage for demo (will be replaced with proper state management)
generation_tasks = {}


@app.get("/")
async def root():
    """Root endpoint."""
    return {"message": "Legal AI Engine Running", "version": "0.1.0"}


@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {"status": "ok"}


@app.post("/api/v1/documents/generate", response_model=GenerateDocumentResponse)
async def generate_document(request: GenerateDocumentRequest):
    """Generate a legal document from natural language description.

    This endpoint initiates document generation and returns a task ID.
    Use the /api/v1/documents/status/{task_id} endpoint to check progress.
    """
    task_id = str(uuid.uuid4())

    # TODO: Implement actual document generation with LangGraph
    # For now, store the request and return task ID
    generation_tasks[task_id] = {
        "status": "PROCESSING",
        "request": request.model_dump(),
        "content": None,
        "error": None,
    }

    return GenerateDocumentResponse(
        task_id=task_id,
        status="PROCESSING",
        message="Document generation started",
    )


@app.get("/api/v1/documents/status/{task_id}", response_model=DocumentGenerationStatus)
async def get_document_status(task_id: str):
    """Get the status of a document generation task."""
    if task_id not in generation_tasks:
        raise HTTPException(status_code=404, detail="Task not found")

    task = generation_tasks[task_id]

    # TODO: Replace with actual task status from LangGraph
    # For demo, simulate completion
    if task["status"] == "PROCESSING":
        task["status"] = "COMPLETED"
        task["content"] = "# Mock Legal Document\n\nThis is a placeholder."

    return DocumentGenerationStatus(
        task_id=task_id,
        status=task["status"],
        content=task.get("content"),
        metadata=task.get("request"),
        error=task.get("error"),
    )


@app.post("/api/v1/qa/ask", response_model=AnswerResponse)
async def ask_question(request: AskQuestionRequest):
    """Ask a legal question and receive an answer with citations.

    The AI will provide answers tailored to the specified mode:
    - LAWYER: Detailed, technical legal analysis
    - SIMPLE: Layperson-friendly explanation
    """
    # TODO: Implement actual Q&A with RAG using PydanticAI
    # For now, return a mock response

    mock_answer = f"""Based on your question: "{request.question}"

This is a placeholder answer. The actual implementation will use:
- PydanticAI for structured responses
- RAG (Retrieval Augmented Generation) for accurate citations
- Polish legal database for authoritative sources

Mode: {request.mode}
"""

    return AnswerResponse(
        answer=mock_answer,
        citations=[
            Citation(
                source="Polish Civil Code",
                article="Art. 118",
                url="https://example.com/civil-code",
            )
        ],
        confidence=0.85,
    )


@app.post("/api/v1/search/rulings", response_model=SearchRulingsResponse)
async def search_rulings(request: SearchRulingsRequest):
    """Search for legal rulings and case law.

    Returns relevant court rulings based on the search query and filters.
    """
    # TODO: Implement actual search integration
    # For now, return mock results

    mock_rulings = [
        Ruling(
            id="ruling-001",
            title=f"Mock ruling for: {request.query}",
            court="Supreme Court of Poland",
            date="2024-01-15",
            summary="This is a placeholder ruling summary.",
            url="https://example.com/ruling/001",
            relevance_score=0.92,
        )
    ]

    return SearchRulingsResponse(
        results=mock_rulings[: request.limit],
        total=len(mock_rulings),
        query=request.query,
    )